move migration related tests to one spec file
